:construction_worker: Use prebuilt Dokken images
One of the challenges with Docker is that it's "process" oriented and
generally not recommended to run Docker containers with an init
system as PID 1, but in the case of doing Test Driven Infrastructure,
that's exactly what we want. Fortunately, Docker provides images
pre-configured with SystemD and require a few minor tweaks
to get them working.

See: https://github.com/test-kitchen/kitchen-dokken#process-orientation