:rotating_light: Write test if Redis is running
After installing Redis via the RedisIO cookbook, we want to check if the
Redis server is running. We do this by checking if a process is listening
on the standard 6379 Redis port. We also check if there's a service
enabled and installed to ensure Redis continues running post reboot.

The name of the service is unique to the way the RedisIO cookbook
works.