:white_check_mark: Ensure can write and read
As a database, Redis wouldn't be very useful if you couldn't read and
write to it. Therefore, we'll use the handy built in `redis-cli` CLI
tool and check if we can write a key and retrieve it's value.

We're going to do that using the super useful `command` resource
that allows us to run _any_ CLI command against the system under
test and check the exit status and the stdout and stderr values.