:recycle: "Refactor" by adding Ubuntu image
The beauty with Test Kitchen and a Configuration Management system
like Chef is that we can keep all the code and test the same, but just
add another `platform` in Test Kitchen and start provisioning and
testing Redis against Ubuntu as well as CentOS. This is a huge win!

Go forth and test your infrastructure. Be a super hero! :muscle: